Fixed byte overflow issue that caused LEDscape to support max 256 px per channel.

+++ ledscape.c
@@ -223,8 +223,8 @@ ledscape_close(
 void
 ledscape_set_color(
 	ledscape_frame_t * const frame,
-	uint8_t strip,
-	uint8_t pixel,
+	uint32_t strip,
+	uint32_t pixel,
 	uint8_t r,
 	uint8_t g,
 	uint8_t b
